+++ packages/mds-metrics-service/server/handlers/utils.ts
@@ -15,8 +15,8 @@
  */
 
 import { Timestamp } from '@mds-core/mds-types'
-import { MetricEntityModel } from './orm/entities/metric-entity'
-import { MetricDomainModel } from '../@types'
+import { MetricEntityModel } from '../orm/entities/metric-entity'
+import { MetricDomainModel } from '../../@types'
 
 export const asMetricDomainModel = ({ id, recorded, ...entity }: MetricEntityModel): MetricDomainModel => entity
